streaming_Youtube/requirements.txt
flask
pandas
numpy
scipy
regex
emot
pyvi
joblib
scipy
sklearn
#logging
kafka-python
findspark
pyspark==3.1.2
IPython
apiclient
tensorflow
keras
google-api-python-client
confluent_kafka
openpyxl
